+++ lib/src/flexible_bottom_sheet.dart
@@ -11,11 +11,43 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 import 'package:bottom_sheet/src/flexible_bottom_sheet_header_delegate.dart';
-import 'package:bottom_sheet/src/widgets/flexible_bottom_sheet_scroll_notifier.dart';
-import 'package:bottom_sheet/src/widgets/flexible_draggable_scrollable_sheet.dart';
+import 'package:bottom_sheet/src/widgets/change_insets_detector.dart';
 import 'package:flutter/material.dart';
+
+/// The signature of a method that provides a [BuildContext] and
+/// [ScrollController] for building a widget that may overflow the draggable
+/// [Axis] of the containing FlexibleDraggableScrollSheet.
+///
+/// Users should apply the [scrollController] to a [ScrollView] subclass, such
+/// as a [SingleChildScrollView], [ListView] or [GridView], to have the whole
+/// sheet be draggable.
+///
+/// [bottomSheetOffset] - fractional value of offset.
+typedef FlexibleDraggableScrollableWidgetBuilder = Widget Function(
+  BuildContext context,
+  ScrollController scrollController,
+  double bottomSheetOffset,
+);
+
+/// The signature of the method that provides [BuildContext]
+/// and [bottomSheetOffset] for determining the position of the BottomSheet
+/// relative to the upper border of the screen.
+/// [bottomSheetOffset] - fractional value of offset.
+typedef FlexibleDraggableScrollableHeaderWidgetBuilder = Widget Function(
+  BuildContext context,
+  double bottomSheetOffset,
+);
+
+/// The signature of a method that provides a [BuildContext]
+/// and [bottomSheetOffset] for determining the position of the BottomSheet
+/// relative to the upper border of the screen.
+/// [bottomSheetOffset] - fractional value of offset.
+typedef FlexibleDraggableScrollableWidgetBodyBuilder = SliverChildDelegate
+    Function(
+  BuildContext context,
+  double bottomSheetOffset,
+);
 
 /// Flexible and scrollable bottom sheet.
 ///
@@ -116,268 +148,132 @@ class FlexibleBottomSheet extends StatefulWidget {
 
 class _FlexibleBottomSheetState extends State<FlexibleBottomSheet>
     with SingleTickerProviderStateMixin {
-  final _topOffsetTween = Tween<double>();
+  final _controller = DraggableScrollableController();
 
+  late double initialChildSize = widget.initHeight;
   bool _isClosing = false;
-
-  late AnimationController _animationController;
-  late double _currentAnchor;
-  late FlexibleDraggableScrollableSheetScrollController _controller;
-  late Animation<double> _topTweenAnimation;
-  late VoidCallback _animationListener;
-  late void Function(AnimationStatus) _statusListener;
-
-  bool _isKeyboardOpenedNotified = false;
-  bool _isKeyboardClosedNotified = false;
-
-  double get _currentExtent => _controller.extent.currentExtent;
-
-  List<double> get _screenAnchors => {
-        if (widget.anchors != null) ...widget.anchors!,
-        widget.maxHeight,
-        widget.minHeight,
-        widget.initHeight,
-      }.toList();
-
-  @override
-  void initState() {
-    super.initState();
-
-    _currentAnchor = widget.initHeight;
-
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 200),
-      vsync: this,
-    );
-    final curve = CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.linear,
-    );
-    _topTweenAnimation = _topOffsetTween.animate(curve);
-    _statusListener = (status) {
-      if (status == AnimationStatus.completed) {
-        _controller.extent.currentExtent = _currentAnchor;
-        _animationController.reset();
-      }
-    };
-    _animationListener = () {
-      if (_animationController.isAnimating) {
-        _controller.extent.currentExtent = _topTweenAnimation.value;
-      }
-    };
-    _topTweenAnimation
-      ..addListener(_animationListener)
-      ..addStatusListener(_statusListener);
-  }
 
   @override
   Widget build(BuildContext context) {
-    _checkKeyboard();
-
-    return FlexibleScrollNotifier(
-      scrollStartCallback: _startScroll,
-      scrollCallback: _scrolling,
-      scrollEndCallback: _endScroll,
-      child: FlexibleDraggableScrollableSheet(
+    return NotificationListener<DraggableScrollableNotification>(
+      onNotification: _scrolling,
+      child: DraggableScrollableSheet(
         maxChildSize: widget.maxHeight,
         minChildSize: widget.minHeight,
-        initialChildSize: widget.initHeight,
+        initialChildSize: initialChildSize,
+        snap: widget.anchors != null,
+        controller: _controller,
+        snapSizes: widget.anchors,
+        expand: widget.isExpand,
         builder: (
           context,
           controller,
         ) {
-          _controller =
-              controller as FlexibleDraggableScrollableSheetScrollController;
-
-          return AnimatedPadding(
-            duration: const Duration(milliseconds: 100),
-            padding: EdgeInsets.only(
-              bottom: MediaQuery.of(context).viewInsets.bottom,
-            ),
-            child: _Content(
-              builder: widget.builder,
-              controller: _controller,
-              decoration: widget.decoration,
-              bodyBuilder: widget.bodyBuilder,
-              headerBuilder: widget.headerBuilder,
-              minHeaderHeight: widget.minHeaderHeight,
-              maxHeaderHeight: widget.maxHeaderHeight,
-              currentExtent: _currentExtent,
+          return ChangeInsetsDetector(
+            handler: (delta, inset) {
+              if (delta > 0) {
+                _animateToFocused(controller);
+                _animateToMaxHeigt();
+              }
+              // checking for openness of the keyboard before opening the sheet
+              if (delta == 0 && inset != 0) {
+                WidgetsBinding.instance!.addPostFrameCallback(
+                  (_) {
+                    setState(() {
+                      initialChildSize = widget.maxHeight;
+                    });
+                  },
+                );
+              }
+            },
+            child: AnimatedPadding(
+              padding: EdgeInsets.only(
+                bottom: MediaQuery.of(context).viewInsets.bottom,
+              ),
+              duration: const Duration(milliseconds: 200),
+              curve: Curves.ease,
+              onEnd: () {
+                _animateToFocused(controller);
+                _updateState();
+              },
+              child: _Content(
+                builder: widget.builder,
+                decoration: widget.decoration,
+                bodyBuilder: widget.bodyBuilder,
+                headerBuilder: widget.headerBuilder,
+                minHeaderHeight: widget.minHeaderHeight,
+                maxHeaderHeight: widget.maxHeaderHeight,
+                currentExtent: _controller.size,
+                scrollController: controller,
+              ),
             ),
           );
         },
-        expand: widget.isExpand,
       ),
     );
   }
 
-  @override
-  void dispose() {
-    _animationController.dispose();
-    _topTweenAnimation
-      ..removeListener(_animationListener)
-      ..removeStatusListener(_statusListener);
-
-    super.dispose();
+  /// Update state
+  /// explicit application [initialChildSize]
+  void _updateState() {
+    setState(() {});
   }
 
-  void _checkKeyboard() {
-    if (MediaQuery.of(context).viewInsets.bottom != 0) {
-      if (!_isKeyboardOpenedNotified) {
-        _isKeyboardOpenedNotified = true;
-        _isKeyboardClosedNotified = false;
-        _keyboardOpened();
-      }
-    } else {
-      if (_isKeyboardOpenedNotified && !_isKeyboardClosedNotified) {
-        _isKeyboardClosedNotified = true;
-        _isKeyboardOpenedNotified = false;
-        _keyboardClosed();
-      }
-    }
-  }
-
-  void _keyboardOpened() {
-    final maxBottomSheetHeight = widget.maxHeight;
-
-    _currentAnchor = maxBottomSheetHeight;
-    _animateToNextAnchor(maxBottomSheetHeight);
-    if (_currentAnchor == widget.maxHeight) {
-      _preScroll();
-    }
-  }
-
-  void _keyboardClosed() {}
-
-  void _preScroll() {
-    if (FocusManager.instance.primaryFocus == null) return;
-
-    final keyboardHeight = MediaQuery.of(context).viewInsets.bottom;
-
-    if (_controller.extent.currentExtent == widget.maxHeight &&
-        keyboardHeight != 0) {
-      final widgetOffset = FocusManager.instance.primaryFocus!.offset.dy;
-      final widgetHeight = FocusManager.instance.primaryFocus!.size.height;
-      final screenHeight = MediaQuery.of(context).size.height;
-
-      final valueToScroll =
-          keyboardHeight - (screenHeight - (widgetOffset + widgetHeight));
-      if (valueToScroll > 0) {
-        Future<void>.delayed(const Duration(milliseconds: 100)).then((_) {
-          _controller.animateTo(
-            _controller.offset + valueToScroll + 10,
-            duration: const Duration(milliseconds: 200),
-            curve: Curves.linear,
-          );
-        });
-      }
-    }
-  }
-
-  bool _startScroll(ScrollStartNotification notification) {
-    return false;
-  }
-
-  bool _scrolling(FlexibleDraggableScrollableNotification notification) {
+  /// Method will be called when scrolling
+  bool _scrolling(DraggableScrollableNotification notification) {
     if (_isClosing) return false;
 
+    initialChildSize = notification.extent;
+
+    _checkNeedCloseBottomSheet(notification.extent);
+
+    return false;
+  }
+
+  /// Make bottom sheet max height
+  void _animateToMaxHeigt() {
+    final currPosition = _controller.size;
+    if (currPosition != widget.maxHeight) {
+      _controller.animateTo(
+        widget.maxHeight,
+        duration: const Duration(milliseconds: 300),
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  /// scroll to focused widget
+  void _animateToFocused(ScrollController controller) {
+    if (FocusManager.instance.primaryFocus == null || _isClosing) return;
+
+    final keyboardHeight = MediaQuery.of(context).viewInsets.bottom;
+    final widgetHeight = FocusManager.instance.primaryFocus!.size.height;
+    final widgetOffset = FocusManager.instance.primaryFocus!.offset.dy;
+    final screenHeight = MediaQuery.of(context).size.height;
+
+    final targetWidgetOffset =
+        screenHeight - keyboardHeight - widgetHeight - 20;
+    final valueToScroll = widgetOffset - targetWidgetOffset;
+    final currentOffset = controller.offset;
+
+    if (valueToScroll > 0) {
+      WidgetsBinding.instance!.addPostFrameCallback((_) {
+        controller.animateTo(
+          currentOffset + valueToScroll,
+          duration: const Duration(milliseconds: 200),
+          curve: Curves.ease,
+        );
+      });
+    }
+  }
+
+  /// checking if the bottom sheet needs to be closed
+  void _checkNeedCloseBottomSheet(double extent) {
     if (widget.isCollapsible && !_isClosing) {
-      if (notification.extent == widget.minHeight) {
-        setState(() {
-          _isClosing = true;
-          _dismiss();
-        });
+      if (extent - widget.minHeight <= 0.005) {
+        _isClosing = true;
+        _dismiss();
       }
-    }
-
-    final currentVal = notification.extent;
-    var initVal = notification.initialExtent;
-
-    if (initVal == 0) {
-      initVal = currentVal;
-    }
-
-    _checkNeedCloseBottomSheet();
-
-    return false;
-  }
-
-  bool _endScroll(ScrollEndNotification notification) {
-    if (widget.anchors?.isNotEmpty ?? false) {
-      _scrollToNearestAnchor(notification.dragDetails);
-    }
-
-    _checkNeedCloseBottomSheet();
-
-    return false;
-  }
-
-  void _checkNeedCloseBottomSheet() {
-    if (_controller.extent.currentExtent <= widget.minHeight && !_isClosing) {
-      _isClosing = true;
-      _dismiss();
-    }
-  }
-
-  void _scrollToNearestAnchor(DragEndDetails? oldDragDetails) {
-    final screenAnchors = _screenAnchors;
-    final nextAnchor = _calculateNextAnchor(screenAnchors);
-
-    _animateToNextAnchor(nextAnchor);
-    _currentAnchor = nextAnchor;
-  }
-
-  double _calculateNextAnchor(List<double> screenAnchor) {
-    if (screenAnchor.contains(_controller.extent.currentExtent)) {
-      return _controller.extent.currentExtent;
-    }
-
-    final nearestAnchor =
-        _findNearestAnchors(screenAnchor, _controller.extent.currentExtent);
-    final firstAnchor = nearestAnchor[0];
-    final secondAnchor = nearestAnchor[1];
-
-    if (firstAnchor == _currentAnchor) {
-      return _findNextAnchorFromPrevious(firstAnchor, secondAnchor);
-    } else if (secondAnchor == _currentAnchor) {
-      return _findNextAnchorFromPrevious(secondAnchor, firstAnchor);
-    } else {
-      return (firstAnchor - _controller.extent.currentExtent).abs() >
-              (secondAnchor - _controller.extent.currentExtent).abs()
-          ? secondAnchor
-          : firstAnchor;
-    }
-  }
-
-  List<double> _findNearestAnchors(List<double> list, double x) {
-    list.sort();
-    final diff = {for (var d in list) d: d - x};
-    final firstAnchor = diff.entries.where((me) => me.value > 0).first.key;
-    final secondAnchor = diff.entries.where((me) => me.value < 0).last.key;
-
-    return [firstAnchor, secondAnchor];
-  }
-
-  double _findNextAnchorFromPrevious(double previousAnchor, double nextAnchor) {
-    const percent = 0.2;
-
-    return (_controller.extent.currentExtent - previousAnchor).abs() >
-            (nextAnchor - previousAnchor).abs() * percent
-        ? nextAnchor
-        : previousAnchor;
-  }
-
-  Future<void> _animateToNextAnchor(double nextAnchor) async {
-    if (_controller.extent.currentExtent == nextAnchor) return;
-
-    _topOffsetTween
-      ..begin = _controller.extent.currentExtent
-      ..end = nextAnchor;
-
-    await _animationController.forward();
-
-    if (nextAnchor <= widget.minHeight && !_isClosing) {
-      _dismiss();
     }
   }
 
@@ -392,17 +288,17 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet>
 /// Content for [FlexibleBottomSheet].
 class _Content extends StatelessWidget {
   final FlexibleDraggableScrollableWidgetBuilder? builder;
-  final FlexibleDraggableScrollableSheetScrollController controller;
   final Decoration? decoration;
   final FlexibleDraggableScrollableHeaderWidgetBuilder? headerBuilder;
   final FlexibleDraggableScrollableWidgetBodyBuilder? bodyBuilder;
   final double? minHeaderHeight;
   final double? maxHeaderHeight;
   final double currentExtent;
+  final ScrollController scrollController;
 
   const _Content({
-    required this.controller,
     required this.currentExtent,
+    required this.scrollController,
     this.builder,
     this.decoration,
     this.headerBuilder,
@@ -417,8 +313,8 @@ class _Content extends StatelessWidget {
     if (builder != null) {
       return builder!(
         context,
-        controller,
-        controller.extent.currentExtent,
+        scrollController,
+        currentExtent,
       );
     }
 
@@ -427,7 +323,7 @@ class _Content extends StatelessWidget {
       child: Container(
         decoration: decoration,
         child: CustomScrollView(
-          controller: controller,
+          controller: scrollController,
           slivers: <Widget>[
             if (headerBuilder != null)
               SliverPersistentHeader(